open mail links in _self target
Non web-based protocol links do not need to be opened in a new window as
they will be handled elsewhere. In fact, opening them in a new window
will result in opening a "void" browser window before letting the proper
application handle it (i.e. the e-mail software in this case).